# PERF: Sparse IntIndex.make_union / Numeric ops