--- a/asv_bench/benchmarks/sparse.py
+++ b/asv_bench/benchmarks/sparse.py
@@ -53,3 +53,42 @@ class sparse_series_to_coo(object):
 
     def time_sparse_series_to_coo(self):
         self.ss.to_coo(row_levels=[0, 1], column_levels=[2, 3], sort_labels=True)
+
+
+class sparse_arithmetic(object):
+    goal_time = 0.2
+
+    def setup(self):
+        np.random.seed(1)
+        self.a_10percent = self.make_sparse_array(length=1000000, dense_size=100000, fill_value=np.nan)
+        self.b_10percent = self.make_sparse_array(length=1000000, dense_size=100000, fill_value=np.nan)
+
+        self.a_10percent_zero = self.make_sparse_array(length=1000000, dense_size=100000, fill_value=0)
+        self.b_10percent_zero = self.make_sparse_array(length=1000000, dense_size=100000, fill_value=0)
+
+        self.a_1percent = self.make_sparse_array(length=1000000, dense_size=10000, fill_value=np.nan)
+        self.b_1percent = self.make_sparse_array(length=1000000, dense_size=10000, fill_value=np.nan)
+
+    def make_sparse_array(self, length, dense_size, fill_value):
+        arr = np.array([fill_value] * length, dtype=np.float64)
+        indexer = np.unique(np.random.randint(0, length, dense_size))
+        arr[indexer] = np.random.randint(0, 100, len(indexer))
+        return pd.SparseArray(arr, fill_value=fill_value)
+
+    def time_sparse_addition_10percent(self):
+        self.a_10percent + self.b_10percent
+
+    def time_sparse_addition_10percent_zero(self):
+        self.a_10percent_zero + self.b_10percent_zero
+
+    def time_sparse_addition_1percent(self):
+        self.a_1percent + self.b_1percent
+
+    def time_sparse_division_10percent(self):
+        self.a_10percent / self.b_10percent
+
+    def time_sparse_division_10percent_zero(self):
+        self.a_10percent_zero / self.b_10percent_zero
+
+    def time_sparse_division_1percent(self):
+        self.a_1percent / self.b_1percent