@jreback 2016-04-30
this looks fine for 0.18.1. do you want to add some sparse benchmarks?